# openapi.json
{
  "components": {
    "schemas": {
      "ApprovalDenied": {
        "additionalProperties": false,
        "description": "사람이 거부했다. 사유는 거부에만 있다. 거부는 실행을 끝내지 않는다.",
        "properties": {
          "approver": {
            "title": "Approver",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "type": {
            "const": "approval_denied",
            "default": "approval_denied",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "ts",
          "type",
          "approver",
          "reason"
        ],
        "title": "ApprovalDenied",
        "type": "object"
      },
      "ApprovalGranted": {
        "additionalProperties": false,
        "description": "사람이 일시정지한 도구 호출을 허가했다. 승인자는 필수다(ADR 0009).",
        "properties": {
          "approver": {
            "title": "Approver",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "type": {
            "const": "approval_granted",
            "default": "approval_granted",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "ts",
          "type",
          "approver"
        ],
        "title": "ApprovalGranted",
        "type": "object"
      },
...
      "ConversationSummarized": {
        "additionalProperties": false,
        "description": "런타임이 앞 요약과 오래된 교환을 새 대화 요약으로 접었다. 에이전트가 내면 실행이 실패한다.",
        "properties": {
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "last_covered_run": {
            "title": "Last Covered Run",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "type": {
            "const": "conversation_summarized",
            "default": "conversation_summarized",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "ts",
          "type",
          "summary",
          "last_covered_run",
          "model",
          "input_tokens",
          "output_tokens"
        ],
        "title": "ConversationSummarized",
        "type": "object"
      },
      "Decision": {
        "discriminator": {
          "mapping": {
            "approve": "#/components/schemas/Approve",
            "deny": "#/components/schemas/Deny"
          },
          "propertyName": "decision"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/Approve"
          },
          {
            "$ref": "#/components/schemas/Deny"
          }
        ]
      },
...
      "ErrorCode": {
        "description": "에러 봉투의 어휘. 상태 코드와 1:1 이다.",
        "enum": [
          "unauthorized",
          "invalid_request",
          "not_found",
          "conflict",
          "internal_error"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "에러 응답의 모양. 성공 응답은 이것으로 감싸지 않는다.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/Violation"
            },
            "title": "Violations",
            "type": "array"
          }
        },
        "required": [
          "code",
          "message",
          "violations",
          "request_id"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "Event": {
        "discriminator": {
          "mapping": {
            "approval_denied": "#/components/schemas/ApprovalDenied",
            "approval_granted": "#/components/schemas/ApprovalGranted",
            "conversation_summarized": "#/components/schemas/ConversationSummarized",
            "llm_called": "#/components/schemas/LlmCalled",
            "run_failed": "#/components/schemas/RunFailed",
            "run_finished": "#/components/schemas/RunFinished",
            "run_paused": "#/components/schemas/RunPaused",
            "run_resumed": "#/components/schemas/RunResumed",
            "run_started": "#/components/schemas/RunStarted",
            "tool_called": "#/components/schemas/ToolCalled"
          },
          "propertyName": "type"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/RunStarted"
          },
          {
            "$ref": "#/components/schemas/LlmCalled"
          },
          {
            "$ref": "#/components/schemas/ToolCalled"
          },
          {
            "$ref": "#/components/schemas/RunPaused"
          },
          {
            "$ref": "#/components/schemas/ApprovalGranted"
          },
          {
            "$ref": "#/components/schemas/ApprovalDenied"
          },
          {
            "$ref": "#/components/schemas/RunResumed"
          },
          {
            "$ref": "#/components/schemas/RunFinished"
          },
          {
            "$ref": "#/components/schemas/RunFailed"
          },
          {
            "$ref": "#/components/schemas/ConversationSummarized"
          }
        ]
      },
...
      "LlmCalled": {
        "additionalProperties": false,
        "description": "모델 호출 하나. 관찰용이면서 재개의 입력이다. prompt 는 루프의 첫 턴에만 담긴다.",
        "properties": {
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "prompt": {
            "default": "",
            "title": "Prompt",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          },
          "tool_calls": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ToolCall"
            },
            "title": "Tool Calls",
            "type": "array"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "type": {
            "const": "llm_called",
            "default": "llm_called",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "ts",
          "type",
          "model",
          "input_tokens",
          "output_tokens",
          "prompt",
          "text",
          "tool_calls"
        ],
        "title": "LlmCalled",
        "type": "object"
      },
...
      "RunFailed": {
        "additionalProperties": false,
        "description": "실행이 실패로 끝났다. error 는 실패의 내용이다. 런타임이 낸다.",
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "type": {
            "const": "run_failed",
            "default": "run_failed",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "ts",
          "type",
          "error"
        ],
        "title": "RunFailed",
        "type": "object"
      },
      "RunFinished": {
        "additionalProperties": false,
        "description": "실행이 출력을 내고 끝났다. 에이전트가 마지막에 하나 낸다.",
        "properties": {
          "output": {
            "title": "Output",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "type": {
            "const": "run_finished",
            "default": "run_finished",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "ts",
          "type",
          "output"
        ],
        "title": "RunFinished",
        "type": "object"
      },
      "RunPaused": {
        "additionalProperties": false,
        "description": "승인을 기다리며 멈췄다. 실패도 끝난 것도 아니다. 도구 하나만 담아 승인과 실행이 1:1이다.",
        "properties": {
          "args": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Json"
            },
            "title": "Args",
            "type": "object"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "type": {
            "const": "run_paused",
            "default": "run_paused",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "ts",
          "type",
          "tool",
          "args"
        ],
        "title": "RunPaused",
        "type": "object"
      },
      "RunResumed": {
        "additionalProperties": false,
        "description": "재생이 끝나고 실제 실행이 다시 시작된다. 재생 구간과 실제 구간의 경계다.",
        "properties": {
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "type": {
            "const": "run_resumed",
            "default": "run_resumed",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "ts",
          "type"
        ],
        "title": "RunResumed",
        "type": "object"
      },
...
      "StartRun": {
        "additionalProperties": false,
        "description": "실행 하나를 일으키는 요청. 등록된 에이전트의 이름과 요청 문자열이다.",
        "properties": {
          "agent": {
            "pattern": "^[a-z][a-z0-9-]{1,62}$",
            "title": "Agent",
            "type": "string"
          },
          "request": {
            "title": "Request",
            "type": "string"
          }
        },
        "required": [
          "agent",
          "request"
        ],
        "title": "StartRun",
        "type": "object"
      },
...
      "Trace": {
        "additionalProperties": false,
        "description": "한 실행의 트레이스. 이벤트는 쓴 순서 그대로다.",
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/TraceEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "run_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "$ref": "#/components/schemas/TraceSchemaVersion"
          }
        },
        "required": [
          "run_id",
          "schema_version",
          "events"
        ],
        "title": "Trace",
        "type": "object"
      },
...
      "TraceSchemaVersion": {
        "enum": [
          "1",
          "2",
          "3"
        ],
        "type": "string"
      },
...
      "Violation": {
        "additionalProperties": false,
        "description": "형식 오류 하나. field 는 `query.limit` 처럼 점으로 이은 경로 문자열이다.",
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "Violation",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "관리는 등록된 플러그인과 지나간 실행을 읽고 플러그인을 켜고 끄며 실행을 일으키지 않는다. 채널은 실행을 일으켜 그 이벤트 스트림을 돌려준다. 두 면은 서로 다른 토큰으로 연다.",
    "title": "Agent OS HTTP API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/runs": {
      "post": {
        "operationId": "start_run",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRun"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "properties": {
                    "data": {
                      "contentMediaType": "application/json",
                      "contentSchema": {
                        "$ref": "#/components/schemas/Event"
                      },
                      "type": "string"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "minimum": 0,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "토큰이 없거나 틀리다"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "찾는 것이 없다"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "있지만 상태나 주체가 요청을 허락하지 않는다"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "요청의 형식이 올바르지 않다"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "서버가 요청을 처리하지 못했다"
          }
        },
        "summary": "실행 하나를 일으켜 그 이벤트를 생기는 대로 흘린다"
      }
    },
    "/runs/{run_id}/approval": {
      "post": {
        "operationId": "decide_approval",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Decision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "properties": {
                    "data": {
                      "contentMediaType": "application/json",
                      "contentSchema": {
                        "$ref": "#/components/schemas/Event"
                      },
                      "type": "string"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "minimum": 0,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "토큰이 없거나 틀리다"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "찾는 것이 없다"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "있지만 상태나 주체가 요청을 허락하지 않는다"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "요청의 형식이 올바르지 않다"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "서버가 요청을 처리하지 못했다"
          }
        },
        "summary": "멈춘 실행에 결정 하나를 내고 재개된 실행의 이벤트를 생기는 대로 흘린다"
      }
    },
    "/runs/{run_id}/continuation": {
      "post": {
        "operationId": "continue_run",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRun"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "properties": {
                    "data": {
                      "contentMediaType": "application/json",
                      "contentSchema": {
                        "$ref": "#/components/schemas/Event"
                      },
                      "type": "string"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "minimum": 0,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "토큰이 없거나 틀리다"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "찾는 것이 없다"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "있지만 상태나 주체가 요청을 허락하지 않는다"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "요청의 형식이 올바르지 않다"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "서버가 요청을 처리하지 못했다"
          }
        },
        "summary": "끝난 실행을 이어 가는 새 실행을 일으켜 그 이벤트를 생기는 대로 흘린다"
      }
    },
...
    "/traces/{run_id}": {
      "get": {
        "operationId": "read_trace",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trace"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "토큰이 없거나 틀리다"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "찾는 것이 없다"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "요청의 형식이 올바르지 않다"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "서버가 요청을 처리하지 못했다"
          }
        },
        "summary": "한 실행의 이벤트 전부를 쓴 순서 그대로"
      }
    }
  }
}